Implement the basics of the game structure
Implement the basic structure of the game loop, making it relatively
easy to expand.

Now the game has one in motion piece at a time that can be pseudo-
randomly generated. A piece will only stop if it hits botton or if
it collides with static pieces (to-be implemented).

This structure assumes that the static pieces of the game will be
stored in memory at an abstraction of a matrix (also to-be
implemented).