Implement all keys using keyboard input
Now it is possible to change the position/rotation of a piece
using aswd.

Furthermore, to implement the input handling for different changes
it was implemented a generic procedure called cycle, that updates
a piece position/rotation, checks for collisions and plots the
negative of the old piece info and the positive of the new one.